+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/ui/screen/home/viewmodel/WalletViewModel.kt
@@ -174,7 +174,7 @@ class WalletViewModel(
                     application = application,
                     zcashNetwork = network,
                     endpoint = LightWalletEndpoint.defaultForNetwork(network),
-                    walletInitMode = WalletInitMode.NewWallet
+                    walletInitMode = WalletInitMode.NewWallet,
                 )
             persistWallet(newWallet)
         }