+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/integration/TestnetIntegrationTest.kt
@@ -165,6 +165,7 @@ class TestnetIntegrationTest : ScopedTest() {
                     // Using existing wallet init mode as simplification for the test
                     walletInitMode = WalletInitMode.ExistingWallet,
                     zcashNetwork = ZcashNetwork.Testnet,
+                    isTorEnabled = false
                 )
         }
     }