+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [WalletCoordinator] now takes [isTorEnabled] as a constructor parameter.
+  - When set to `true`, lightwalletd RPC queries will be made over Tor (where possible and beneficial).
+  - When set to `false`, lightwalletd RPC queries will always be made directly to the server.
+- [Synchronizer] now exposes [initializationError] property containing synchronizer errors that happened during 
+  synchronizer init
+
 ### Fixed
 - Tor client is now optional in case it's instantiation fails to prevent SDK